outgoing docs
